+++ Scripts/Utils/MapLabelUtil.cs
@@ -19,8 +19,9 @@ public static class MapLabelUtil
 	/// Adds a child to this parent with a <see cref="TMP_Text"/> component
 	/// </summary>
 	/// <param name="parent">A transform that can be seen on the radar</param>
+	/// <param name="labeloffset">The offset for the label with respect to the mapCamera</param>
 	/// <param name="setRotationInUpdate">Should a <see cref="RotateWithMapCamera"/> component be added to this object?</param>
-	public static TMP_Text AddLabelObject(GameObject parent, bool setRotationInUpdate = true)
+	public static TMP_Text AddLabelObject(GameObject parent, Vector2 labeloffset, bool setRotationInUpdate = true)
 	{
 		GameObject labelObject = new GameObject(LABEL_OBJECT_NAME);
 		Transform labelObjectTransform = labelObject.transform;
@@ -30,6 +31,8 @@ public static class MapLabelUtil
 		labelObjectTransform.localPosition = labelPosition;
 		labelObjectTransform.rotation      = MapCameraRotationObserver.MapCameraRotation;
 		labelObjectTransform.localScale    = labelScale;
+		
+		labelObject.AddComponent<LabelOffset>().Offset = labeloffset;
 
 		// Prevent non-uniform scaling in the parent
 		Vector3 parentScale = parent.transform.localScale;